pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.8.1"
uvicorn = "~0.23.2"
fastapi = "~0.101.0"
python-multipart = "~0.0.6"
email-validator = "^2.0.0.post2"
requests = "~2.31.0"
passlib = { version = "~1.7.4", extras = ["bycrypt"] }
tenacity = "~8.2.2"
pydantic = "~2.1.1"
emails = "~0.6"
raven = "~6.10.0"
gunicorn = "~21.2.0"
psycopg2-binary = "~2.9.7"
alembic = "~1.11.2"
sqlalchemy = "~2.0.19"
python-jose = { version = "~3.3.0", extras = ["cryptography"] }
setuptools = "~68.0.0"
pydantic-settings = "~2.0.2"
flaat = "^1.1.14"
gitpython = "^3.1.32"

[tool.poetry.group.dev.dependencies]
pytest = "~7.4.0"
pytest-postgresql = "~5.0.0"
pytest-xdist = "~3.3.1"
pytest-cov = "~4.1.0"
bandit = "~1.7.5"
mypy = "~1.5.0"
black = "~23.7.0"
isort = "~5.12.0"
autoflake = "~2.2.0"
flake8 = "~6.1.0"
sqlalchemy-stubs = "~0.4"
httpx = "~0.24.1"

[tool.pylint]
max-line-length=120
ignored-modules = ["alembic"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[build-system]
requires = ["poetry>=1.5.1"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "--import-mode=importlib"
